+++ hbw/app/javascript/packs/hbw/components/claiming/short_list.jsx
@@ -1,13 +1,25 @@
+/* eslint-disable camelcase */
 import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 import { parseISO, isPast } from 'date-fns';
+import DueDate from './shared/due_date';
+import CreatedDate from './shared/created_date';
 
 const HBWClaimingShortList = ({ env, tasks }) => {
   const { translator: t, connection, taskListPath } = env;
 
   const viewAll = () => { window.location.href = `${connection.options.host}/${taskListPath}`; };
   const goToTask = (entityUrl) => { window.location.href = entityUrl; };
+
+  const renderDate = (created, due) => {
+    if (due) {
+      return <span className="deadline"><DueDate env={env} dateISO={due} /></span>;
+    } else {
+      return <span><CreatedDate env={env} dateISO={created} /></span>;
+    }
+  };
+
   const rowCN = due => (
     cx({
       row:     true,
@@ -18,15 +30,15 @@ const HBWClaimingShortList = ({ env, tasks }) => {
   return (
     <div className="short-list">
       {tasks.map(({
-        id, icon, name, due, created, entityUrl
+        id, icon, name, due, created, entity_url
       }) => (
-        <div onClick={() => goToTask(entityUrl)} key={id} className={rowCN(due)}>
+        <div onClick={() => goToTask(entity_url)} key={id} className={rowCN(due)}>
           <div className="left">
             <i className={icon} />
             <span className="title">{name}</span>
           </div>
           <div className="right">
-            <span className="created">{due || created}</span>
+            {renderDate(created, due)}
           </div>
         </div>
       ))}
@@ -42,12 +54,12 @@ const HBWClaimingShortList = ({ env, tasks }) => {
 HBWClaimingShortList.propTypes = {
   env:   PropTypes.object.isRequired,
   tasks: PropTypes.arrayOf(PropTypes.shape({
-    id:        PropTypes.string.isRequired,
-    name:      PropTypes.string.isRequired,
-    icon:      PropTypes.string.isRequired,
-    entityUrl: PropTypes.string.isRequired,
-    created:   PropTypes.string.isRequired,
-    due:       PropTypes.string,
+    id:         PropTypes.string.isRequired,
+    name:       PropTypes.string.isRequired,
+    icon:       PropTypes.string.isRequired,
+    entity_url: PropTypes.string.isRequired,
+    created:    PropTypes.string.isRequired,
+    due:        PropTypes.string,
   })).isRequired,
 };