+++ CHANGELOG.md
@@ -66,6 +66,7 @@ v1.8.0 [unreleased]
 - [#346](https://github.com/latera/homs/pull/346) Add code coverage for backend.
 - [#352](https://github.com/latera/homs/pull/352) Wrap task list page together.
 - [#351](https://github.com/latera/homs/pull/351) Improve claim button styles on widget form.
+- [#354](https://github.com/latera/homs/pull/354) Wrap task short pop up list page together.
 
 v1.7.8 [2019-10-03]
 -------------------